requirements.txt
# ============================================================
#  Factor Arbitrage RMT — Requirements
#  Python >= 3.11
# ============================================================

# ---------- Core Científico ----------
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.4.0
statsmodels>=0.14.0

# ---------- Datos de Mercado ----------
yfinance>=0.2.38

# ---------- Serialización ----------
pyarrow>=16.0.0

# ---------- Visualización ----------
matplotlib>=3.8.0
seaborn>=0.13.0

# ---------- Configuración ----------
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ---------- Utilidades ----------
tqdm>=4.66.0
loguru>=0.7.2
click>=8.1.7

# ---------- Testing ----------
pytest>=8.0.0
pytest-cov>=5.0.0
